+++ payment/internal/server/authorize_card.go
@@ -82,7 +82,7 @@ func (s *Server) AuthorizeCard(ctx context.Context, in *paymentproto.AuthorizeCa
 
 	return &paymentproto.AuthorizeCardResponse{
 		PaymentMethod: &paymentproto.PaymentMethod{
-			Id:             int64(paymentMethod.Id),
+			Id:             paymentMethod.Id,
 			LastFourDigits: paymentMethod.LastFourDigits,
 		},
 	}, nil
